+++ src/Tests/Infrastructure/New-VaultProject.Tests.ps1
@@ -27,7 +27,7 @@ Describe 'Vault Infra Tests' {
         New-Item -Path $outPutPath -ItemType Directory  -ErrorAction SilentlyContinue
     } #beforeAll
 
-    Context 'Module Checks' {
+    Context 'Vault Checks' {
 
         BeforeEach {
             Remove-Item -Path $outPutPathStar -Recurse -Force
@@ -365,6 +365,89 @@ Describe 'Vault Infra Tests' {
 
                     $azureModuleFiles.Name.Contains('azure-pipelines.yml') | Should -BeExactly $true
                     $azureModuleFiles.Name.Contains('actions_bootstrap.ps1') | Should -BeExactly $true
+                    $azureModuleFiles.Name.Contains('pull_request_template.md') | Should -BeExactly $false
+
+                    $installContentPath = [System.IO.Path]::Combine($outPutPath, 'actions_bootstrap.ps1')
+                    $installContent = Get-Content -Path $installContentPath -Raw
+                    $installContent | Should -BeLike '*Microsoft.PowerShell.SecretManagement*'
+
+                    $azureYMLContentPath = [System.IO.Path]::Combine($outPutPath, 'azure-pipelines.yml')
+                    $azureYMLContent = Get-Content -Path $azureYMLContentPath -Raw
+                    $azureYMLContent | Should -BeLike "*build_ps_WinLatest*"
+                    $azureYMLContent | Should -BeLike "*build_pwsh_WinLatest*"
+                    $azureYMLContent | Should -BeLike "*build_pwsh_ubuntuLatest*"
+                    $azureYMLContent | Should -BeLike "*build_pwsh_macOSLatest*"
+                } #it
+
+                It 'should generate an Azure Pipelines based module stored on Azure Repos with all required elements' {
+                    $vaultParameters = @{
+                        VAULT        = 'text'
+                        ModuleName   = 'SecretManagement.MyVault'
+                        Description  = 'text'
+                        Version      = '0.0.1'
+                        FN           = 'user full name'
+                        CICD         = 'AZURE'
+                        AzureOptions = 'windows', 'pwshcore', 'linux', 'macos'
+                        RepoType     = 'AZURE'
+                        License      = 'None'
+                        Changelog    = 'NOCHANGELOG'
+                        COC          = 'NOCONDUCT'
+                        Contribute   = 'NOCONTRIBUTING'
+                        Security     = 'NOSECURITY'
+                        CodingStyle  = 'Stroustrup'
+                        Pester       = '5'
+                        PassThru     = $true
+                        NoLogo       = $true
+                    }
+                    $eval = New-VaultProject -VaultParameters $vaultParameters -DestinationPath $outPutPath
+                    $eval | Should -Not -BeNullOrEmpty
+
+                    $azureModuleFiles = Get-ChildItem -Path $outPutPathStar -Recurse -Force
+
+                    $azureModuleFiles.Name.Contains('azure-pipelines.yml') | Should -BeExactly $true
+                    $azureModuleFiles.Name.Contains('actions_bootstrap.ps1') | Should -BeExactly $true
+                    $azureModuleFiles.Name.Contains('pull_request_template.md') | Should -BeExactly $true
+
+                    $installContentPath = [System.IO.Path]::Combine($outPutPath, 'actions_bootstrap.ps1')
+                    $installContent = Get-Content -Path $installContentPath -Raw
+                    $installContent | Should -BeLike '*Microsoft.PowerShell.SecretManagement*'
+
+                    $azureYMLContentPath = [System.IO.Path]::Combine($outPutPath, 'azure-pipelines.yml')
+                    $azureYMLContent = Get-Content -Path $azureYMLContentPath -Raw
+                    $azureYMLContent | Should -BeLike "*build_ps_WinLatest*"
+                    $azureYMLContent | Should -BeLike "*build_pwsh_WinLatest*"
+                    $azureYMLContent | Should -BeLike "*build_pwsh_ubuntuLatest*"
+                    $azureYMLContent | Should -BeLike "*build_pwsh_macOSLatest*"
+                } #it
+
+                It 'should generate an Azure Pipelines based module stored on Bitbucket with all required elements' {
+                    $vaultParameters = @{
+                        VAULT        = 'text'
+                        ModuleName   = 'SecretManagement.MyVault'
+                        Description  = 'text'
+                        Version      = '0.0.1'
+                        FN           = 'user full name'
+                        CICD         = 'AZURE'
+                        AzureOptions = 'windows', 'pwshcore', 'linux', 'macos'
+                        RepoType     = 'BITBUCKET'
+                        License      = 'None'
+                        Changelog    = 'NOCHANGELOG'
+                        COC          = 'NOCONDUCT'
+                        Contribute   = 'NOCONTRIBUTING'
+                        Security     = 'NOSECURITY'
+                        CodingStyle  = 'Stroustrup'
+                        Pester       = '5'
+                        PassThru     = $true
+                        NoLogo       = $true
+                    }
+                    $eval = New-VaultProject -VaultParameters $vaultParameters -DestinationPath $outPutPath
+                    $eval | Should -Not -BeNullOrEmpty
+
+                    $azureModuleFiles = Get-ChildItem -Path $outPutPathStar -Recurse -Force
+
+                    $azureModuleFiles.Name.Contains('azure-pipelines.yml') | Should -BeExactly $true
+                    $azureModuleFiles.Name.Contains('actions_bootstrap.ps1') | Should -BeExactly $true
+                    $azureModuleFiles.Name.Contains('pull_request_template.md') | Should -BeExactly $false
 
                     $installContentPath = [System.IO.Path]::Combine($outPutPath, 'actions_bootstrap.ps1')
                     $installContent = Get-Content -Path $installContentPath -Raw